+++ webtraversallibrary/js/dom.js
@@ -718,5 +718,3 @@ const eventMap = {
       }
     }
   };
-
-